Split lib code again for easier development
